+++ packages/core/src/index.ts
@@ -3,6 +3,9 @@ export * from './verify.module.js';
 export * from './verify.service.js';
 export * from './controllers/verify.controller.js';
 export * from './controllers/dto.js';
+export * from './controllers/swagger/verify.swagger.js';
+export * from './errors.js';
+export * from './utils/async-handler.js';
 
 export * from './store/memory-verify.store.js';
 export * from './store/memory-abuse.store.js';